% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_c.R
\name{update_c}
\alias{update_c}
\title{Update Clusters}
\usage{
update_c(mu_1, mu_2, pdat1, pdat2, ncluster)
}
\arguments{
\item{mu_1}{The center values for each cluster in condition 1.}

\item{mu_2}{The center values for each cluster in condition 2.}

\item{pdat1}{The centered data from condition 1, columns should be
samples (cells) and rows should be features (genes).}

\item{pdat2}{The centered data from condition 2, columns should be
samples (cells) and rows should be features (genes). The number of genes
should be the same as \code{pdat1}.
as in pdat1.}

\item{ncluster}{The number of clusters present in the data.}
}
\value{
A list containing the cluster membership for condition 1 and
condition 2.
}
\description{
Updates the cluster membership for each iteration of SparseDC. Runs inside
\code{sparse_dc_fun}.
}
